requirements.txt
PySide6==6.6.0
paddleocr==2.6.1.3
paddlepaddle==2.6.2
opencv-python==4.5.1.48
pillow==10.0.0
numpy==1.24.3
pyqtgraph==0.13.1
pandas==1.5.3
sqlite3  # 通常Python自带